moved to multi-module maven project
created a parent maven project that includes the Frontend and Backend as separated modules.
added the frontend as maven project to use the frontend-maven-plugin to manage the project and integrate it as dependency of the backend.
This will then package the frontend as jar dependency and include it in the backend.jar as static website that can be accessed directly next to the API.

This would allow the Project to be run as an AIO solution instead of separate instances for the frontend and the backend.